requirements.txt
beautifulsoup4==4.9.0
Flask==1.1.2
pandas==1.0.3
PyYAML==5.3.1
requests==2.23.0

SQLAlchemy~=1.3.16
WTForms~=2.3.1
geopy~=1.22.0